pyproject.toml
[project]
name = "audiobox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.19.0",
    "einops>=0.8.0",
    "gradio>=5.8.0",
    "huggingface-hub[cli]>=0.26.1",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.2.34",
    "lightning>=2.4.0",
    "mlflow>=2.17.0",
    "mutagen>=1.47.0",
    "openai>=1.57.0",
    "polars>=1.11.0",
    "psutil>=6.1.0",
    "pynvml>=11.5.3",
    "soundfile>=0.12.1",
    "tensorboard>=2.18.0",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchdiffeq>=0.2.4",
    "torchlibrosa>=0.1.0",
    "transformers>=4.46.0",
    "tritonclient[http]>=2.52.0",
]

[tool.uv]
dev-dependencies = [
    "torch-tb-profiler>=0.4.3",
]
